Migrate Recaptcha integration (#185)
The old library had a race condition. #145 tried to address it.

It’s because the donate island (<GivingFormWithTotals>) was being
rendered/loaded twice.
So the recaptcha provider was being ran twice.
It looked for the recaptcha script tag and if it found it, it would call
`onLoad` immediately.
That was the race, because the script could exist in DOM, but the
browser had not loaded & executed the script. So the `window.recaptcha`
global instance was not available yet.

This fork handles multiple instances correctly, and is more
updated/maintained.
Easy switch & win.